1.阶段提交

+++ app/src/main/java/com/ayla/hotelsaas/ui/GatewayAddActivity.java
@@ -63,7 +63,7 @@ public class GatewayAddActivity extends BaseMvpActivity<GatewayAddGuideView, Gat
         refreshBindShow();
         mPresenter.registerDeviceWithDSN(
                 getIntent().getStringExtra("dsn"),
-                getIntent().getLongExtra("cuId", 0),
+                getIntent().getIntExtra("cuId", 0),
                 getIntent().getLongExtra("scopeId", 0),
                 getIntent().getStringExtra("deviceName"));
     }
